+++ data.php
@@ -1,0 +1,45 @@
+<?php 
+$posts = [
+  [
+      'type' => 'post-quote',
+      'title' => 'Цитата',
+      'post_content' => 'Мы в жизни любим только раз, а после ищем лишь похожих',
+      'author' => 'Лариса',
+      'avatar' => 'userpic-larisa-small.jpg'
+  ],
+  [
+      'type' => 'post-text',
+      'title' => 'Игра престолов',
+      'post_content' => 'Не могу дождаться начала финального сезона своего любимого сериала!',
+      'author' => 'Владик',
+      'avatar' => 'userpic.jpg'
+  ],
+  [
+      'type' => 'post-photo',
+      'title' => 'Наконец, обработал фотки!',
+      'post_content' => 'rock-medium.jpg',
+      'author' => 'Виктор',
+      'avatar' => 'userpic-mark.jpg'
+  ],
+  [
+      'type' => 'post-photo',
+      'title' => 'Моя мечта',
+      'post_content' => 'coast-medium.jpg',
+      'author' => 'Лариса',
+      'avatar' => 'userpic-larisa-small.jpg'
+  ],
+  [
+      'type' => 'post-link',
+      'title' => 'Лучшие курсы',
+      'post_content' => 'www.htmlacademy.ru',
+      'author' => 'Владик',
+      'avatar' => 'userpic.jpg'
+  ],
+  [
+      'type' => 'post-text',
+      'title' => 'Пишем первую функцию',
+      'post_content' => 'Чтобы карточки оставались компактными и не занимали слишком много места размер содержимого надо принудительно ограничивать. Для фотографий и видео это можно сделать через CSS, для цитат и ссылок есть ограничение длины при создании поста. Остаётся текстовый контент. Его длина никак не ограничивается в момент создания, а т. к. пользователи могут писать очень длинные тексты, необходимо предусмотреть обрезание текста до приемлемой длины при показе карточки поста на странице популярного.',
+      'author' => 'Владик',
+      'avatar' => 'userpic.jpg'
+  ]
+];
