+++ index.php
@@ -1,23 +1,6 @@
 <?php
-
-$posts = [];
-
-function include_template($name, $data) {
-    $name = 'templates/' . $name;
-    $result = '';
-
-    if (!file_exists($name)) {
-        return $result;
-    }
-
-    ob_start();
-    extract($data);
-    require $name;
-
-    $result = ob_get_clean();
-
-    return $result;
-}
+require_once('functions.php');
+require_once('data.php');
 
 $page_content = include_template('main.php', ['posts' => $posts]);
 
@@ -28,5 +11,4 @@ $layout_content = include_template('layout.php', [
 
 print($layout_content);
 
-?>
